fixed bug #2
Still wasn't going through all links on updates. Fixed that issue by an
accidental [j] indexing. Created a function that would be call to check
if link contains new episode. So that it can be called from the
notifications since they are now created in background, they couldn't
call the runtime mesage, to update it.